Update Integer32.cs
ToErrorCode() should not throw exceptions as the method is invoked by error message display and ToString() methods. Also, as the ErrorCode field of a SNMP packet is read/parsed by this class, a device migth brake/create a DOS by simply sending an error code unknown to the lib. We'd loose compatibility and introduce issues without any need or gain.